Move/rename files to follow naming conventions
