+++ activerecord/test/cases/fixture_set/file_test.rb
